refactor: Parser함수를 Adapter로 분리
