+++ src/shared/api/server/service/symbol.service.ts
@@ -5,15 +5,7 @@ import {
 } from "@/shared/api/server/provider/api-provider";
 import type { FinnhubSymbolResult, Symbol } from "@/shared/domain/symbol";
 
-/** Finnhub 응답을 도메인 심볼 리스트로 변환 */
-export const toSymbols = (payload: FinnhubSymbolResult): Symbol[] => {
-  return payload.result.map((item) => ({
-    description: item.description,
-    displaySymbol: item.displaySymbol,
-    symbol: item.symbol,
-    type: item.type,
-  }));
-};
+import { toSymbols } from "../adapter/symbol.adapter";
 
 /** 공급자 심볼 검색 */
 export const searchSymbols = async (
